feat(config): unify runtime config health checks
