+++ tests/test_setup.sh
@@ -116,8 +116,11 @@ checks = [
     data.get("u16", {}).get("limit") == 800,
     data.get("circuit_breaker", {}).get("threshold") == 3,
     data.get("circuit_breaker", {}).get("cooldown_seconds") == 300,
+    data.get("circuit_breaker", {}).get("lock_timeout_seconds") == 5,
     data.get("w14", {}).get("cooldown_seconds") == 3600,
     data.get("paralysis", {}).get("threshold") == 7,
+    data.get("write_escalate_threshold") == 5,
+    data.get("learn", {}).get("metrics_tail_bytes") == 5242880,
 ]
 raise SystemExit(0 if all(checks) else 1)
 PY
